fix(mount): Align cfg file to regular command line
This commit enables the parsing of the following formats of options
written in the cfg file (if provided, obviusly): <-D value>-like
format, this is a single character keyword like -P, -D, -B or -S and
<-o opt1=val1,opt2=val2,...> format.

Plus, it enables the random casing in the names of the options and
forbids mount if cfg is provided and there are some unexpected (like
--help, -V, -u options) or wrong options. This commit also adds an
integration test that check the expected random casing parsing in
cfg files.

Aside from those changes, a couple of comments were added to
sfs_opt_parse_cfg_file function to ease its reading.